worker/requirements.txt
redis>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
celery>=5.3.0
boto3>=1.34.0
google-cloud-vision>=3.5.0
